selection of rule subset by picking out least redundant rule.

+++ source/main.py
@@ -2,8 +2,10 @@ import time, copy, os, pickle, glob, csv, ast
 from config import parameters
 from PatternHandler import PatternHandler
 from DependencyGraphHandler import DependencyGraphHandler
+from SubsetHandler import SubsetHandler
 import pandas as pd
 from collections import defaultdict
+from sklearn.feature_selection import mutual_info_classif
 from tqdm import tqdm
 tqdm.pandas()
 from pandarallel import pandarallel
@@ -20,6 +22,8 @@ output_pattern_counter_pkl_filepath = parameters.output_pattern_counter_pkl_file
 output_targets_dir = parameters.output_targets_dir
 output_targets_concat_csv_filepath = parameters.output_targets_concat_csv_filepath
 output_pattern_evaluation_csv_filepath = parameters.output_pattern_evaluation_csv_filepath
+output_subset_selection_log_filepath = parameters.output_subset_selection_log_filepath
+output_final_report_csv_filepath = parameters.output_final_report_csv_filepath
 
 def match_opinion_words(content, opinion_word_lexicon):
     opinion_words = []
@@ -124,15 +128,75 @@ def pattern_quality_estimation(domain, original_df, pattern_counter, pattern_han
     
     print('Evaluating rules for [%s]..' % domain)
     filepath = output_pattern_evaluation_csv_filepath % domain
-    f = open(filepath, 'w', encoding='utf-8-sig')
-    wr = csv.writer(f)  
-    wr.writerow(['domain', 'pattern', 'count', 'precision_multiple', 'recall_multiple', 'f1_multiple', 'precision_distinct', 'recall_distinct', 'f1_distinct'])
-    for pattern in concat_df['pattern'].unique():
-        current_df = concat_df[concat_df['pattern']==pattern]
-        pre_mul, rec_mul, pre_dis, rec_dis = calculate_precision_recall(current_df)
-        wr.writerow([domain, pattern, '%d'%current_df['pattern_count'].iloc[0], '%.2f'%pre_mul, '%.2f'%rec_mul, '%.2f'%calculate_f1(pre_mul,rec_mul), '%.2f'%pre_dis, '%.2f'%rec_dis, '%.2f'%calculate_f1(pre_dis,rec_dis)])
-    f.close()
+    if not os.path.exists(filepath): 
+        f = open(filepath, 'w', encoding='utf-8-sig')
+        wr = csv.writer(f)  
+        wr.writerow(['domain', 'pattern', 'count', 'precision_multiple', 'recall_multiple', 'f1_multiple', 'precision_distinct', 'recall_distinct', 'f1_distinct'])
+        for pattern in concat_df['pattern'].unique():
+            current_df = concat_df[concat_df['pattern']==pattern]
+            pre_mul, rec_mul, pre_dis, rec_dis = calculate_precision_recall(current_df)
+            wr.writerow([domain, pattern, '%d'%current_df['pattern_count'].iloc[0], '%.2f'%pre_mul, '%.2f'%rec_mul, '%.2f'%calculate_f1(pre_mul,rec_mul), '%.2f'%pre_dis, '%.2f'%rec_dis, '%.2f'%calculate_f1(pre_dis,rec_dis)])
+        f.close()
+        print('Created %s' % filepath)
+    pattern_evaluation_df = pd.read_csv(filepath)
+    print('Loaded %s' % filepath)
+    return concat_df, pattern_evaluation_df
+
+def evaluate_rule_set(original_df, selected_pattern_list, pattern_handler, dependency_handler):
+    df = copy.deepcopy(original_df)
+    df['predicted_targets'] = df.apply(lambda x: list(), axis=1)
+    for one_flattened_dep_rels in selected_pattern_list:
+        dep_rels = one_flattened_dep_rels.split('-')
+        df['predicted_targets'] = df.apply(lambda x: pattern_handler.extract_targets(x['doc'], x['opinion_words'], dep_rels, dependency_handler, x['predicted_targets']), axis=1)
+
+    pre_mul, rec_mul, pre_dis, rec_dis = calculate_precision_recall(df)
+    f1_mul = calculate_f1(pre_mul,rec_mul)
+    f1_dis = calculate_f1(pre_dis,rec_dis)
+    return f1_mul, f1_dis
+
+def pick_least_redundant_one_pattern(selected_pattern_list, subset_handler):
+    min_redundancy_degree_score, min_redundant_pattern = 999, None
+    x1 = subset_handler.evaluate_patterns_tp(selected_pattern_list)['tp'].values.reshape(-1,1)
+    for candidate_pattern in subset_handler.pattern_list:
+        if candidate_pattern in selected_pattern_list: continue
+        patterns2 = copy.deepcopy(selected_pattern_list)
+        patterns2.append(candidate_pattern)
+        x2 = subset_handler.evaluate_patterns_tp(patterns2)['tp'].values.reshape(-1,1)
+        redundancy_degree_score = mutual_info_classif(x1, x2, discrete_features=[0])
+        print('%s=>%.4f' % (candidate_pattern, redundancy_degree_score))
+        if redundancy_degree_score < min_redundancy_degree_score:
+            min_redundancy_degree_score = redundancy_degree_score
+            min_redundant_pattern = candidate_pattern
+    return min_redundant_pattern, min_redundancy_degree_score
+
+def pattern_subset_selection(domain, original_df, subset_handler, pattern_handler, dependency_handler):
+    print('Processing subset selection for [%s]..' % domain)
+    
+    filepath = output_subset_selection_log_filepath % domain
+    text_file = open(filepath, "w", encoding='utf-8')
+    
+    best_f1_mul, best_f1_dis, best_subset = 0, 0, []
+    selected_pattern_list = [subset_handler.pattern_list[0]]
+    f1_mul, f1_dis = evaluate_rule_set(original_df, selected_pattern_list, pattern_handler, dependency_handler)
+    content = "Selected pattern list = %s \n\tF1 (multiple): %.4f\tF1 (distinct): %.4f" % (str(selected_pattern_list), f1_mul, f1_dis)
+    print("[%s]Selected pattern list = %s \n\tF1 (multiple): %.4f\tF1 (distinct): %.4f" % (domain, str(selected_pattern_list), f1_mul, f1_dis))
+    while f1_mul > best_f1_mul:
+        best_f1_mul, best_f1_dis, best_subset = f1_mul, f1_dis, copy.deepcopy(selected_pattern_list)
+
+        min_redundant_pattern, mi_score = pick_least_redundant_one_pattern(selected_pattern_list, subset_handler)
+        content += "\nLeast redundant pattern = %s [Redundancy score (MI score) = %.4f]" % (min_redundant_pattern, mi_score)
+        print("[%s]Least redundant pattern = %s [Redundancy score (MI score) = %.4f]" % (domain, min_redundant_pattern, mi_score))
+        selected_pattern_list.append(min_redundant_pattern)
+        f1_mul, f1_dis = evaluate_rule_set(original_df, selected_pattern_list, pattern_handler, dependency_handler)
+        content += "\n\nSelected pattern list = %s \n\tF1 (multiple): %.4f\tF1 (distinct): %.4f" % (str(selected_pattern_list), f1_mul, f1_dis)
+        print("[%s]Selected pattern list = %s \n\tF1 (multiple): %.4f\tF1 (distinct): %.4f" % (domain, str(selected_pattern_list), f1_mul, f1_dis))
+        
+    content += "\n\n<Best> Selected pattern list = %s \n\tF1 (multiple): %.4f\tF1 (distinct): %.4f" % (str(best_subset), best_f1_mul, best_f1_dis)
+    print("[%s]<Best> Selected pattern list = %s \n\tF1 (multiple): %.4f\tF1 (distinct): %.4f" % (domain, str(best_subset), best_f1_mul, best_f1_dis))
+    text_file.write(content)
+    text_file.close()
     print('Created %s' % filepath)
+    return best_f1_mul, best_f1_dis, best_subset
 
 def save_pkl(item_to_save, filepath):
     with open(filepath, 'wb') as f:
@@ -162,6 +226,9 @@ def main():
         raw_df['targets'] = raw_df.progress_apply(lambda x: pattern_handler.process_targets(x['content'], x['raw_targets']), axis=1) 
         save_pkl(raw_df, output_raw_df_pkl_filepath)
     
+    f = open(output_final_report_csv_filepath, 'w', encoding='utf-8-sig')
+    wr = csv.writer(f)  
+    wr.writerow(['Domain', 'Measure', 'All', 'Best subset'])
     for domain in raw_df['domain'].unique():
         print('Processing [%s]..' % domain)
         df = raw_df[raw_df['domain']==domain]
@@ -172,7 +239,19 @@ def main():
             pattern_counter = pattern_extraction(domain, df, pattern_handler, dependency_handler)
             save_pkl(pattern_counter, filepath)
         
-        pattern_quality_estimation(domain, df, pattern_counter, pattern_handler, dependency_handler)
+        predicted_targets_df, pattern_evaluation_df = pattern_quality_estimation(domain, df, pattern_counter, pattern_handler, dependency_handler)
+        
+        print('Calculating true positives for each sentence..')
+        predicted_targets_df['tp'] = predicted_targets_df.progress_apply(lambda row: calculate_true_positive(row['predicted_targets'], row['targets']), axis=1)
+        
+        subset_handler = SubsetHandler(domain, predicted_targets_df, pattern_evaluation_df)
+        best_f1_mul, best_f1_dis, best_subset = pattern_subset_selection(domain, df, subset_handler, pattern_handler, dependency_handler)
+        
+        all_f1_mul, all_f1_dis = evaluate_rule_set(df, subset_handler.pattern_list, pattern_handler, dependency_handler)
+        wr.writerow([domain, 'F1 score (multiple)', '%.4f'%all_f1_mul, '%.4f'%best_f1_mul])
+        wr.writerow([domain, 'F1 score (distinct)', '%.4f'%all_f1_dis, '%.4f'%best_f1_dis])
+    f.close()
+    print('Created %s' % output_final_report_csv_filepath)
     
 def elapsed_time(start):
     end = time.time()
